requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.0
cvzone==1.6
cycler==0.12.1
Django==5.0.3
djangorestframework==3.14.0
filelock==3.13.1
flatbuffers==24.3.7
fonttools==4.50.0
fsspec==2024.3.0
gdown==4.6.0
hydra-core==1.3.2
idna==3.6
jax==0.4.25
jaxlib==0.4.25
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
mediapipe==0.10.11
ml-dtypes==0.3.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.1
pillow==10.2.0
protobuf==3.20.3
psutil==5.9.8
py-cpuinfo==9.0.0
pycparser==2.21
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sounddevice==0.4.6
soupsieve==2.5
sqlparse==0.4.4
sympy==1.12
thop==0.1.1.post2209072238
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.10.0
tzdata==2024.1
ultralytics==8.1.29
urllib3==2.2.1
